Add specs for Float.to_human_size and Fixnum.to_human_size
this is simply encoding the current outputs. I'm not sure whether they
should be considered _correct_ at this point.

Wire up the specs via the Rakefile and TravisCI configuration.

Fix the development dependencies: the Rakefile depends on rake and
jeweler. RSpec requires an update to work with newer rake; since it's
not actually used yet I removed the version constraint.

Bundle update to re-resolve all dependencies.