Add support for retrieving a single pool's addresses in asset find
Currently, the only way to retrieve the addresses for an asset is to get
all addresses associated with an asset.  It's pretty common for us to
have multiple addresses per asset (e.g. management network, application
network) so it's a minor hassle for us to have to parse it.  This allows
for a special tag of `addresses_POOLNAME`, where `POOLNAME` (obviously) is
an ip address pool.  Any addresses in that pool are then passed into
`format_asset_value` and thereby formatted like the `addresses` tag is.